authorization policies added

+++ RestaurantAPI ASP.NET Core/Services/RestaurantService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI_ASP.NET_Core.Authorization;
 using RestaurantAPI_ASP.NET_Core.Entities;
 using RestaurantAPI_ASP.NET_Core.Exceptions;
 using RestaurantAPI_ASP.NET_Core.Model;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RestaurantAPI_ASP.NET_Core.Services
@@ -16,12 +19,16 @@ namespace RestaurantAPI_ASP.NET_Core.Services
         private readonly RestaurantDbContext  _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public RestaurantDto GetById(int id)
         {
@@ -49,6 +56,7 @@ namespace RestaurantAPI_ASP.NET_Core.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
@@ -62,6 +70,13 @@ namespace RestaurantAPI_ASP.NET_Core.Services
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
 
+            var authorizationresult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationresult.Succeeded)
+            {
+                throw new ForbidException("Forbidden");
+            }
+
             if (restaurant == null) throw new NotFoundException("Not found to delete");
 
             _dbContext.Restaurants.Remove(restaurant);
@@ -70,13 +85,21 @@ namespace RestaurantAPI_ASP.NET_Core.Services
 
         public void Update(UpdateRestaurantDto dto, int id)
         {
+
             var restaurant = _dbContext
                 .Restaurants
                 .FirstOrDefault(r => r.Id == id);
 
             if (restaurant == null) throw new NotFoundException("Not found to update");
 
-            if(dto.Name != null)
+            var authorizationresult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationresult.Succeeded)
+            {
+                throw new ForbidException("Forbidden");
+            }
+
+            if (dto.Name != null)
                 restaurant.Name = dto.Name;
             
             if (dto.Description != null)
